server/requirements.txt
flask==3.0.0
google-generativeai==0.3.2
requests==2.31.0
pillow==12.0.0
beautifulsoup4==4.12.2
pytest==8.2.2
python-dotenv==1.0.0
